requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
supabase==2.0.0
python-dotenv==1.0.0
matplotlib==3.8.2